% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{a0<-}
\alias{a0<-}
\title{Create generic \verb{a0<-} function for class}
\usage{
a0(x) <- value
}
\arguments{
\item{x}{An \code{Mlr} object.}

\item{value}{Numeric shape parameter for residual variance prior to assign to
slot.}
}
\value{
None.
}
\description{
Create generic \verb{a0<-} function for class
}
\examples{
m1 <- Mlr(ndocs = 1)
a0(m1) <- 1.0
}
