% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{a0}
\alias{a0}
\title{Create generic \code{a0} function for class}
\usage{
a0(x)
}
\arguments{
\item{x}{An \code{Mlr} object.}
}
\value{
Double value of shape prior parameter for residual variance.
}
\description{
Create generic \code{a0} function for class
}
\examples{
m1 <- Mlr(ndocs = 1)
a0(m1)
}
