% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{b0<-}
\alias{b0<-}
\title{Create generic \verb{b0<-} function for class}
\usage{
b0(x) <- value
}
\arguments{
\item{x}{An \code{Mlr} object.}

\item{value}{Numeric value of rate parameter for residual variance prior to
assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{b0<-} function for class
}
\examples{
m1 <- Mlr(ndocs = 1)
b0(m1) <- 1.0
}
