% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{b0}
\alias{b0}
\title{Create generic \code{b0} function for class}
\usage{
b0(x)
}
\arguments{
\item{x}{An \code{Mlr} object.}
}
\value{
Double value of rate prior parameter for residual variance.
}
\description{
Create generic \code{b0} function for class
}
\examples{
m1 <- Mlr(ndocs = 1)
b0(m1)
}
