% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_logical}
\alias{check_logical}
\title{Check for logical and non-missing argument}
\usage{
check_logical(arg)
}
\arguments{
\item{arg}{Argument to check.}
}
\description{
Check for logical and non-missing argument
}
