% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_topic_word_cpp}
\alias{count_topic_word_cpp}
\title{Count topic-word co-occurences in corpus (ntopic x nvocab) (parallelizable)}
\usage{
count_topic_word_cpp(D, K, V, doc_topic, doc_word)
}
\arguments{
\item{D}{The number of documents in the corpus.}

\item{K}{The number of topics.}

\item{V}{The number of terms in the corpus vocabulary.}

\item{doc_topic}{A D x max(\eqn{N_d}) matrix of topic assignments for
the corpus.}

\item{doc_word}{A D x max(\eqn{N_d}) matrix of words for corpus.}
}
\description{
Count topic-word co-occurences in corpus (ntopic x nvocab) (parallelizable)
}
