% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_beta}
\alias{draw_beta}
\alias{.draw_beta}
\title{Sample \eqn{B} from full conditional distribution}
\usage{
.draw_beta(wz_co, gamma_)
}
\arguments{
\item{wz_co}{A K x V matrix of counts of word-topic co-occurrences
(topics: columns; words: rows).}

\item{gamma_}{The hyperparameter on the Dirichlet prior for \eqn{\beta_k}.}

\item{m}{Number of samples}
}
\value{
A K x V matrix \eqn{B}.
}
\description{
Sample \eqn{B} from full conditional distribution
}
