% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_betak}
\alias{draw_betak}
\alias{.draw_betak}
\title{Sample \eqn{\beta_k} from full conditional distribution}
\usage{
.draw_betak(wz_co, gamma_)
}
\arguments{
\item{wz_co}{A V x 1 vector of counts of the draws of each word for topic
k over all documents.}

\item{gamma_}{The hyperparameter for the Dirichlet priors on \eqn{\beta_k}.}
}
\value{
A V x 1 vector of estimates for \eqn{\beta_k}.
}
\description{
Sample \eqn{\beta_k} from full conditional distribution
}
