% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_theta}
\alias{draw_theta}
\alias{.draw_theta}
\title{Sample \eqn{\Theta} from full conditional distribution}
\usage{
.draw_theta(z_count, alpha_)
}
\arguments{
\item{z_count}{A D x K matrix of counts of topic draws (columns) in
documents (rows).}

\item{alpha_}{The hyperparameter on the Dirichlet prior for \eqn{\theta_d}.}
}
\value{
A D x K matrix \eqn{\Theta}.
}
\description{
Sample \eqn{\Theta} from full conditional distribution
}
