% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_thetad}
\alias{draw_thetad}
\alias{.draw_thetad}
\title{Sample \eqn{\theta_d} from full conditional distribution}
\usage{
.draw_thetad(z_count, alpha_)
}
\arguments{
\item{z_count}{A K x 1 vector of counts of topic draw in document d.}

\item{alpha_}{The hyperparameter on the Dirichlet prior for \eqn{\theta_d}.}
}
\value{
A K x 1 vector draw of \eqn{\theta_d}.
}
\description{
Sample \eqn{\theta_d} from full conditional distribution
}
