% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{est_beta,Lda-method}
\alias{est_beta,Lda-method}
\title{Generic function to estimate mean/median beta matrix}
\usage{
\S4method{est_beta}{Lda}(mcmc_fit, docs, burn = 0, thin = 1,
  stat = "mean")
}
\arguments{
\item{mcmc_fit}{An Lda object.}

\item{docs}{A D x max(N_d) matrix of the words in all documents. Unused word
positions should be set to 0.}

\item{burn}{The number of draws to discard as a burn-in period. Default: 0.}

\item{thin}{The number of draws to skip as a thinning period. Default: 1 (no thinning).}

\item{stat}{The summary statistic to use on the posterior draws. Default: \code{"mean"}.}
}
\description{
Generic function to estimate mean/median beta matrix
}
