% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_betak_cpp}
\alias{est_betak_cpp}
\title{Estimate beta_k}
\usage{
est_betak_cpp(k, V, wz_co, gamma_)
}
\arguments{
\item{k}{The topic label (i.e., the row of the K x V beta matrix).}

\item{V}{The number of terms in the corpus vocabulary.}

\item{wz_co}{A V x 1 vector of counts of the draws of each word for topic
\eqn{k} over all documents.}

\item{gamma_}{The hyperparameter for the Dirichlet priors on \eqn{\beta_k}.}
}
\value{
A V x 1 vector of estimates for beta_k.
}
\description{
Estimate beta_k
}
