% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_thetad}
\alias{est_thetad}
\title{Estimate \eqn{\theta_d}}
\usage{
est_thetad(z_count, alpha_)
}
\arguments{
\item{z_count}{A K x 1 vector of counts of topic draw in document d.}

\item{alpha_}{The hyperparameter on the Dirichlet prior for \eqn{\theta_d}.}
}
\value{
A K x 1 vector of estimate for \eqn{\theta_d}.
}
\description{
Estimate \eqn{\theta_d}
}
