% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_thetad_cpp}
\alias{est_thetad_cpp}
\title{Estimate theta_d (on log scale)}
\usage{
est_thetad_cpp(z_count, alpha_, K)
}
\arguments{
\item{z_count}{A K x 1 vector of counts of topic draw in document \eqn{d}.}

\item{alpha_}{The hyperparameter on the Dirichlet prior for \eqn{\theta_d}.}

\item{K}{The number of topics.}
}
\description{
Estimate theta_d (on log scale)
}
