% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{eta}
\alias{eta}
\title{Create generic \code{eta} function for class}
\usage{
eta(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
A numeric matrix of posterior draws of regression coefficients.
}
\description{
Create generic \code{eta} function for class
}
\examples{
m1 <- Model(ndocs = 1)
eta(m1)
}
