% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eta_logpost_glm}
\alias{eta_logpost_glm}
\title{Compute full conditional log-posterior of eta for logistic regression}
\usage{
eta_logpost_glm(y, x, eta, mu0, sigma0)
}
\arguments{
\item{y}{A D x 1 vector of the outcome variable for each document.}

\item{x}{A D x (p + 1) matrix of additional predictors including a column of
1's for the intercept.}

\item{eta}{A (p + 1) x 1 vector of regression coefficients}

\item{mu0}{A (p + 1) x 1 vector of prior means for the regression
coefficients.}

\item{sigma0}{A (p + 1) x (p + 1) prior variance-covariance matrix
for the regression coefficients.}
}
\description{
Compute full conditional log-posterior of eta for logistic regression
}
