% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{eta_start<-}
\alias{eta_start<-}
\title{Create generic \verb{eta_start<-} function for class}
\usage{
eta_start(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Numeric vector of starting values for regression coefficients to
assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{eta_start<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
eta_start(m1) <- rep(0.0, times = 2)

}
