% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{eta_start}
\alias{eta_start}
\title{Create generic \code{eta_start} function for class}
\usage{
eta_start(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
Numeric vector of starting values for regression coefficients.
}
\description{
Create generic \code{eta_start} function for class
}
\examples{
m1 <- Model(ndocs = 1)
eta_start(m1)
}
