% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{extra<-}
\alias{extra<-}
\title{Create generic \verb{extra<-} function for class}
\usage{
extra(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{List of additional model fitting information to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{extra<-} function for class
}
