% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{extra}
\alias{extra}
\title{Create generic \code{extra} function for class}
\usage{
extra(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
A list of model fitting information including time elapsed, label
switching correction status, and the original function call.
}
\description{
Create generic \code{extra} function for class
}
\examples{
m1 <- Model(ndocs = 1)
extra(m1)
}
