% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{get_toptopics,Lda-method}
\alias{get_toptopics,Lda-method}
\title{Compute the most probable topics in a document}
\usage{
\S4method{get_toptopics}{Lda}(mcmc_fit, ntopics, burn = 0, thin = 1,
  stat = "mean")
}
\arguments{
\item{mcmc_fit}{An Lda object.}

\item{ntopics}{The number of topics to retrieve.}

\item{burn}{The number of draws to discard as a burn-in period. Default: 0.}

\item{thin}{The number of draws to skip as a thinning period. Default: 1 (no thinning).}

\item{stat}{The summary statistic to use on the posterior draws. Default: \code{"mean"}.}
}
\description{
Compute the most probable topics in a document
}
