% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{get_toptopics,Lda,numeric-method}
\alias{get_toptopics,Lda,numeric-method}
\title{Return most probable words for topic}
\usage{
\S4method{get_toptopics}{Lda,numeric}(mcmc_fit, n_topics, burn = 0,
  thin = 1, stat = "mean")
}
\description{
Return most probable words for topic
}
