% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{get_toptopics,Lda,numeric,numeric,character-method}
\alias{get_toptopics,Lda,numeric,numeric,character-method}
\title{Return most probable topics for each document.}
\usage{
\S4method{get_toptopics}{Lda,numeric,numeric,character}(mcmc_fit,
  burn = 0, thin = 1, stat = "mean")
}
\value{
A tibble with three columns: \code{doc}: the document number,
  \code{topic}: the topic number, \code{prob}: the probability of each topic.
}
\description{
Return most probable topics for each document.
}
