% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{get_topwords,Lda,numeric,character-method}
\alias{get_topwords,Lda,numeric,character-method}
\title{Return most probable words for each topic.}
\usage{
\S4method{get_topwords}{Lda,numeric,character}(mcmc_fit, nwords, vocab,
  burn = 0, thin = 1, method = "termscore", stat = "mean")
}
\arguments{
\item{mcmc_fit}{An Lda object.}
}
\value{
A tibble with three columns: \code{topic}: the topic number,
  \code{word}: the vocabulary term, \code{prob}: the term-score or
  probability of each word for a given topic.
}
\description{
Return most probable words for each topic.
}
