% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{get_topwords,matrix,numeric,character-method}
\alias{get_topwords,matrix,numeric,character-method}
\title{Return most probable words for each topic.}
\usage{
\S4method{get_topwords}{matrix,numeric,character}(beta_, nwords, vocab,
  method = "termscore")
}
\arguments{
\item{beta_}{A K x V matrix of topic-word probabilities.}
}
\value{
A tibble with three columns: \code{topic}: the topic number,
  \code{word}: the vocabulary term, \code{prob}: the term-score or
  probability of each word for a given topic.
}
\description{
Return most probable words for each topic.
}
