% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-generic-functions.R
\name{get_zbar}
\alias{get_zbar}
\title{Generic function to retrieve the empirical topic proportions}
\usage{
get_zbar(mcmc_fit, burn = 0, thin = 1)
}
\arguments{
\item{mcmc_fit}{An Lda object.}

\item{burn}{The number of draws to discard as a burn-in period. Default: 0.}

\item{thin}{The number of draws to skip as a thinning period. Default: 1 (no thinning).}
}
\description{
Compute empirical topic proportions (zbar) from \code{@topics} in an Lda
object
}
