% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{gg_coef,Sldalogit-method}
\alias{gg_coef,Sldalogit-method}
\title{Generic function to plot the regression coefficients for sLDAX models}
\usage{
\S4method{gg_coef}{Sldalogit}(mcmc_fit, beta_, nwords, vocab, varnames,
  burn = 0, thin = 1, method = "termscore", stat = "mean",
  errorbw = 0.5)
}
\arguments{
\item{mcmc_fit}{An Sldalogit object.}

\item{beta_}{A \eqn{K} x \eqn{V} matrix of \eqn{V} vocabulary probabilities
for each of \eqn{K} topics.}

\item{nwords}{The number of words to retrieve.}

\item{vocab}{A character vector containing the vocabulary.}

\item{varnames}{A character vector of variable names for additional
predictors (if any).}

\item{burn}{The number of draws to discard as a burn-in period. Default: 0.}

\item{thin}{The number of draws to skip as a thinning period. Default: 1 (no thinning).}

\item{method}{If "termscore", use term scores (similar to tf-idf). If "prob",
use probabilities. Default: "termscore".}

\item{stat}{The summary statistic to use on the posterior draws. Default: \code{"mean"}.}

\item{errorbw}{Controls the width of the +/- 2 posterior standard error bars.}
}
\description{
Generic function to plot the regression coefficients for sLDAX models
}
