% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_lda}
\alias{gibbs_lda}
\title{Collapsed Gibbs sampler for the LDA model}
\usage{
gibbs_lda(m, burn, docs, w, K, alpha_ = 0.1, gamma_ = 1.01,
  display_progress = FALSE)
}
\arguments{
\item{m}{The number of iterations to run the Gibbs sampler.}

\item{burn}{The number of iterations to discard as the burn-in period.}

\item{docs}{A D x max(\eqn{N_d}) matrix of word indices for all documents.}

\item{w}{A D x V matrix of counts for all documents and vocabulary terms.}

\item{K}{The number of topics.}

\item{alpha_}{The hyper-parameter for the prior on the topic proportions
(default: 0.1).}

\item{gamma_}{The hyper-parameter for the prior on the topic-specific
vocabulary probabilities (default: 1.01).}

\item{display_progress}{Should percent progress of sampler be displayed
(default: \code{FALSE}).}
}
\description{
Collapsed Gibbs sampler for the LDA model
}
