% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gibbs_mlr}
\alias{gibbs_mlr}
\title{Fit linear regression model}
\usage{
gibbs_mlr(formula, data, m = 100, burn = 0, y = NULL, x = NULL,
  mu0 = NULL, sigma0 = NULL, a0 = NULL, b0 = NULL,
  eta_start = NULL, verbose = FALSE, display_progress = FALSE)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}}: a symbolic
description of the model to be fitted.}

\item{data}{An optional data frame containing the variables in the model.}

\item{m}{The number of iterations to run the Gibbs sampler (default: 100).}

\item{burn}{The number of iterations to discard as the burn-in period
(default: 0).}

\item{y}{An optional D x 1 vector of binary outcomes (0/1) to be predicted.
Not needed if \code{data} is supplied.}

\item{x}{An optional D x p design matrix of additional predictors (do NOT
include an intercept column!). Not needed if \code{data} is supplied.}

\item{mu0}{An optional p x 1 mean vector for the prior on the regression
coefficients. See 'Details'.}

\item{sigma0}{A p x p variance-covariance matrix for the prior on the
regression coefficients. See 'Details'.}

\item{a0}{The shape parameter for the prior on sigma2 (default: 0.001).}

\item{b0}{The scale parameter for the prior on sigma2 (default: 0.001).}

\item{eta_start}{A p x 1 vector of starting values for the regression
coefficients.}

\item{verbose}{Should parameter draws be output during sampling? (default:
\code{FALSE}).}

\item{display_progress}{Should percent progress of sampler be displayed
(default: \code{FALSE}). Recommended that only one of \code{verbose} and
\code{display_progress} be set to \code{TRUE} at any given time.}
}
\value{
An object of class \code{\linkS4class{Mlr}}.
}
\description{
\code{gibbs_mlr} is used to fit a Bayesian linear regression model using
Gibbs sampling.
}
\details{
For \code{mu0}, by default, we use a vector of \eqn{p} 0s for \eqn{p}
regression coefficients.

For \code{sigma0}, by default, we use a \eqn{p} x \eqn{p} identity matrix.
}
\seealso{
Other Gibbs sampler: \code{\link{gibbs_logistic_cpp}},
  \code{\link{gibbs_logistic}},
  \code{\link{gibbs_mlr_cpp}}, \code{\link{gibbs_sldax}}
}
\concept{Gibbs sampler}
