% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_mlr_cpp}
\alias{gibbs_mlr_cpp}
\alias{.gibbs_mlr_cpp}
\title{Collapsed Gibbs sampler for multiple linear regression}
\usage{
.gibbs_mlr_cpp(
  m,
  burn,
  thin,
  y,
  x,
  mu0,
  sigma0,
  eta_start,
  a0 = 0.001,
  b0 = 0.001,
  verbose = FALSE,
  display_progress = FALSE
)
}
\arguments{
\item{m}{The number of iterations to run the Gibbs sampler.}

\item{burn}{The number of iterations to discard as the burn-in period.}

\item{thin}{The period of iterations to keep after the burn-in period
(default: \code{1}).}

\item{y}{A D x 1 vector of outcomes to be predicted.}

\item{x}{A D x (p + 1) matrix of additional predictors.}

\item{mu0}{A (p + 1) x 1 mean vector for the prior on the regression
coefficients.}

\item{sigma0}{A (p + 1) x (p + 1) variance-covariance matrix for the
prior on the regression coefficients.}

\item{eta_start}{A (p + 1) x 1 vector of starting values for the
regression coefficients.}

\item{a0}{The shape parameter for the prior on sigma2 (default: \code{0.001})}

\item{b0}{The scale parameter for the prior on sigma2 (default: \code{0.001})}

\item{verbose}{Should parameter draws be output during sampling? (default:
\code{false}).}

\item{display_progress}{Should percent progress of sampler be displayed
(default: \code{false}). Recommended that only one of \code{verbose} and
\code{display_progress} be set to \code{true} at any given time.}
}
\value{
An object of class \code{\link[=Mlr-class]{Mlr}}.
}
\description{
Collapsed Gibbs sampler for multiple linear regression
}
