% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_slda_logit}
\alias{gibbs_slda_logit}
\title{Collapsed Gibbs sampler for the sLDA-X model with a binary outcome}
\usage{
gibbs_slda_logit(m, burn, y, docs, w, K, mu0, sigma0, eta_start,
  proposal_sd, alpha_ = 0.1, gamma_ = 1.01, verbose = FALSE,
  display_progress = FALSE)
}
\arguments{
\item{m}{The number of iterations to run the Gibbs sampler.}

\item{burn}{The number of iterations to discard as the burn-in period.}

\item{y}{A D x 1 vector of binary outcomes (0/1) to be predicted.}

\item{docs}{A D x max(\eqn{N_d}) matrix of word indices for all documents.}

\item{w}{A D x V matrix of counts for all documents and vocabulary terms.}

\item{K}{The number of topics.}

\item{mu0}{A K x 1 mean vector for the prior on the regression coefficients.}

\item{sigma0}{A K x K variance-covariance matrix for the prior on the
regression coefficients.}

\item{eta_start}{A K x 1 vector of starting values for the regression
coefficients.}

\item{proposal_sd}{A vector of proposal standard deviations for drawing the
regression coefficients, N(0, proposal_sd) (default: 0.2).}

\item{alpha_}{The hyper-parameter for the prior on the topic proportions
(default: 0.1).}

\item{gamma_}{The hyper-parameter for the prior on the topic-specific
vocabulary probabilities (default: 1.01).}

\item{verbose}{Should parameter draws be output during sampling? (default:
\code{FALSE}).}

\item{display_progress}{Should percent progress of sampler be displayed
(default: \code{FALSE}). Recommended that only one of \code{verbose} and
\code{display_progress} be set to \code{TRUE} at any given time.}
}
\description{
Collapsed Gibbs sampler for the sLDA-X model with a binary outcome
}
