% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sldax_cpp}
\alias{gibbs_sldax_cpp}
\title{Collapsed Gibbs sampler for the sLDA-X model with a binary outcome}
\usage{
gibbs_sldax_cpp(docs, w, m, burn, K, model, y, x, mu0, sigma0, a0, b0,
  eta_start, proposal_sd, interaction_xcol = -1L, alpha_ = 0.1,
  gamma_ = 1.01, constrain_eta = TRUE, verbose = FALSE,
  display_progress = FALSE)
}
\arguments{
\item{docs}{A D x max(\eqn{N_d}) matrix of word indices for all documents.}

\item{w}{A D x V matrix of counts for all documents and vocabulary terms.}

\item{m}{The number of iterations to run the Gibbs sampler.}

\item{burn}{The number of iterations to discard as the burn-in period.}

\item{K}{The number of topics.}

\item{y}{A D x 1 vector of binary outcomes (0/1) to be predicted.}

\item{x}{A D x p matrix of additional predictors (no column of 1s for
intercept).}

\item{mu0}{A K x 1 mean vector for the prior on the regression coefficients.}

\item{sigma0}{A (K + p + 1) x (K + p + 1) variance-covariance matrix for the
prior on the regression coefficients. The first p + 1 columns/rows
correspond to predictors in X, while the last K columns/rows correspond to
the K topic means.}

\item{eta_start}{A (K + p) x 1 vector of starting values for the
regression coefficients. The first p elements correspond to predictors
in X, while the last K elements correspond to the K topic means.}

\item{proposal_sd}{The proposal standard deviation for drawing the
regression coefficients, N(0, proposal_sd) (default: 0.2).}

\item{alpha_}{The hyper-parameter for the prior on the topic proportions
(default: 0.1).}

\item{gamma_}{The hyper-parameter for the prior on the topic-specific
vocabulary probabilities (default: 1.01).}

\item{constrain_eta}{A logical (default = \code{TRUE}): If \code{TRUE}, the
regression coefficients will be constrained so that they are in descending
order; if \code{FALSE}, no constraints will be applied.}

\item{verbose}{Should parameter draws be output during sampling? (default:
\code{FALSE}).}

\item{display_progress}{Should percent progress of sampler be displayed
(default: \code{FALSE}). Recommended that only one of \code{verbose} and
\code{display_progress} be set to \code{TRUE} at any given time.}
}
\description{
Collapsed Gibbs sampler for the sLDA-X model with a binary outcome
}
