% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.non_negative_integer}
\alias{is.non_negative_integer}
\title{Check for non-negative integer}
\usage{
is.non_negative_integer(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Argument to check.}

\item{tol}{Tolerance (default: sqrt(.Machine$double.eps))}
}
\description{
Check for non-negative integer
}
