% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.positive_integer}
\alias{is.positive_integer}
\title{Check for positive integer}
\usage{
is.positive_integer(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Argument to check.}

\item{tol}{Tolerance (default: sqrt(.Machine$double.eps))}
}
\description{
Check for positive integer
}
