% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.whole_number}
\alias{is.whole_number}
\title{Check for integer argument}
\usage{
is.whole_number(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Argument to check.}

\item{tol}{Tolerance (default: sqrt(.Machine$double.eps))}
}
\description{
Check for integer argument
}
