% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-class.R
\docType{class}
\name{Lda-class}
\alias{Lda-class}
\alias{Lda}
\title{An S4 class to represent a LDA model.}
\description{
An S4 class to represent a LDA model.
}
\section{Slots}{

\describe{
\item{\code{beta}}{A K x V x M numeric array of draws of topic-word probabilities}

\item{\code{theta}}{A D x K x M numeric array of draws of document-topic
probabilities}

\item{\code{alpha}}{A numeric prior hyperparameter for theta}

\item{\code{gamma}}{A numeric prior hyperparameter for beta}
}}

