% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{loglike<-}
\alias{loglike<-}
\title{Create generic \verb{loglike<-} function for class}
\usage{
loglike(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Numeric vector of log likelihoods to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{loglike<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
loglike(m1) <- rep(NaN, times = nchain(m1))
}
