% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{loglike}
\alias{loglike}
\title{Create generic \code{loglike} function for class}
\usage{
loglike(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
A numeric vector of log-likelihood values across sampler iterations.
}
\description{
Create generic \code{loglike} function for class
}
\examples{
m1 <- Model(ndocs = 1)
loglike(m1)

}
