% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{logpost<-}
\alias{logpost<-}
\title{Create generic \verb{logpost<-} function for class}
\usage{
logpost(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Numeric vector of log posteriors to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{logpost<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
logpost(m1) <- rep(NaN, times = nchain(m1))
}
