% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{logpost}
\alias{logpost}
\title{Create generic \code{logpost} function for class}
\usage{
logpost(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
A numeric vector of log-posterior values across sampler iterations.
}
\description{
Create generic \code{logpost} function for class
}
\examples{
m1 <- Model(ndocs = 1)
logpost(m1)
}
