% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{lpd<-}
\alias{lpd<-}
\title{Create generic \verb{lpd<-} function for class}
\usage{
lpd(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Numeric matrix of log predictive densities in each document to
assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{lpd<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
lpd(m1) <- matrix(NaN, nrow = 1, ncol = 1)
}
