% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr-class.R
\name{Mlr}
\alias{Mlr}
\alias{sigma2}
\alias{sigma2,Mlr-method}
\alias{sigma2<-}
\alias{sigma2<-,Mlr-method}
\alias{a0}
\alias{a0,Mlr-method}
\alias{a0<-}
\alias{a0<-,Mlr-method}
\alias{b0}
\alias{b0,Mlr-method}
\alias{b0<-}
\alias{b0<-,Mlr-method}
\title{Mlr class}
\usage{
Mlr(a0 = 0.001, b0 = 0.001, sigma2 = NaN, ...)

sigma2(x)

\S4method{sigma2}{Mlr}(x)

sigma2(x) <- value

\S4method{sigma2}{Mlr}(x) <- value

a0(x)

\S4method{a0}{Mlr}(x)

a0(x) <- value

\S4method{a0}{Mlr}(x) <- value

b0(x)

\S4method{b0}{Mlr}(x)

b0(x) <- value

\S4method{b0}{Mlr}(x) <- value

Mlr(a0 = 0.001, b0 = 0.001, sigma2 = NaN, ...)
}
\description{
S4 class for a regression model. Inherits from \code{\linkS4class{Model}}.
}
\section{Slots}{

\describe{
\item{\code{a0}}{A prior shape hyperparameter for sigma2.}

\item{\code{b0}}{A prior rate hyperparameter for sigma2.}

\item{\code{sigma2}}{A nchain x 1 numeric vector of draws of the residual variance.}
}}

