% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{mu0<-}
\alias{mu0<-}
\title{Create generic \verb{mu0<-} function for class}
\usage{
mu0(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Numeric vector of prior means for regression coefficients to
assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{mu0<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
mu0(m1) <- rep(0.0, times = 2)
}
