% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{mu0}
\alias{mu0}
\title{Create generic \code{mu0} function for class}
\usage{
mu0(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
Numeric vector of prior means for regression coefficients \code{eta}.
}
\description{
Create generic \code{mu0} function for class
}
\examples{
m1 <- Model(ndocs = 1)
mu0(m1)
}
