% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{nchain<-}
\alias{nchain<-}
\title{Create generic \verb{nchain<-} function for class}
\usage{
nchain(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Integer length of sampler chain to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{nchain<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
nchain(m1) <- 100
}
