% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{nchain}
\alias{nchain}
\title{Create generic \code{nchain} function for class}
\usage{
nchain(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
Integer length of sampler chain.
}
\description{
Create generic \code{nchain} function for class
}
\examples{
m1 <- Model(ndocs = 1)
nchain(m1)
}
