% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{ndocs<-}
\alias{ndocs<-}
\title{Create generic \verb{ndocs<-} function for class}
\usage{
ndocs(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Integer number of documents to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{ndocs<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
ndocs(m1) <- 2
}
