% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{ndocs}
\alias{ndocs}
\title{Create generic \code{ndocs} function for class}
\usage{
ndocs(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
Integer number of documents.
}
\description{
Create generic \code{ndocs} function for class
}
\examples{
m1 <- Model(ndocs = 1)
ndocs(m1)
}
