% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{ntopics<-}
\alias{ntopics<-}
\title{Create generic \verb{ntopics<-} function for class}
\usage{
ntopics(x) <- value
}
\arguments{
\item{x}{An \code{Sldax} object.}

\item{value}{Integer number of topics to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{ntopics<-} function for class
}
\examples{
m1 <- Sldax(ndocs = 1, nvocab = 2,
            topics = array(c(1, 2, 2, 1), dim = c(1, 4, 1)),
            theta = array(c(0.5, 0.5), dim = c(1, 2, 1)),
            beta = array(c(0.5, 0.5, 0.5, 0.5), dim = c(2, 2, 1)))
ntopics(m1) <- 2
}
