% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{nvocab}
\alias{nvocab}
\title{Create generic \code{nvocab} function for class}
\usage{
nvocab(x)
}
\arguments{
\item{x}{An \code{Sldax} object.}
}
\value{
Integer number of unique terms in vocabulary.
}
\description{
Create generic \code{nvocab} function for class
}
\examples{
m1 <- Sldax(ndocs = 1, nvocab = 2,
            topics = array(c(1, 2, 2, 1), dim = c(1, 4, 1)),
            theta = array(c(0.5, 0.5), dim = c(1, 2, 1)),
            beta = array(c(0.5, 0.5, 0.5, 0.5), dim = c(2, 2, 1)))
nvocab(m1)
}
