% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{p_eff<-}
\alias{p_eff<-}
\title{Create generic \verb{p_eff<-} function for class}
\usage{
p_eff(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Numeric value of effective number of parameters estimate from
WAIC to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{p_eff<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
p_eff(m1) <- NaN
}
