% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{p_eff}
\alias{p_eff}
\title{Create generic \code{p_eff} function for class}
\usage{
p_eff(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
Numeric estimate of the number of effective parameters when computing
WAIC.
}
\description{
Create generic \code{p_eff} function for class
}
\examples{
m1 <- Model(ndocs = 1)
p_eff(m1)
}
