% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{post_pred_glm}
\alias{post_pred_glm}
\title{Posterior predictive likelihood for logistic regression}
\usage{
post_pred_glm(x, eta)
}
\arguments{
\item{x}{A D x (p + 1) matrix of additional predictors.}

\item{eta}{A (p + 1) x 1 vector of regression coefficients.}
}
\description{
Posterior predictive likelihood for logistic regression
}
