% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{post_pred_sldax_logit}
\alias{post_pred_sldax_logit}
\title{Posterior predictive log-likelihood for sLDA logistic}
\usage{
post_pred_sldax_logit(x, zbar, eta)
}
\arguments{
\item{x}{A D x (p + 1) matrix of additional predictors including a column of
1's for the intercept.}

\item{zbar}{A D x K matrix with row \eqn{d} containing the mean number of
draws of topics \eqn{z_1, \ldots, z_K} in document \eqn{d} where each row
sums to 1.}

\item{eta}{A (p + K) x 1 vector of regression coefficients.}
}
\description{
Posterior predictive log-likelihood for sLDA logistic
}
