% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{proposal_sd<-}
\alias{proposal_sd<-}
\title{Create generic \verb{proposal_sd<-} function for class}
\usage{
proposal_sd(x) <- value
}
\arguments{
\item{x}{An \code{Logistic} object.}

\item{value}{Numeric vector of scale parameters for Metropolis sampling of
regression coefficients to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{proposal_sd<-} function for class
}
\examples{
m1 <- Logistic(ndocs = 1)
proposal_sd(m1) <- c(2.38, 2.38)
}
