% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{proposal_sd}
\alias{proposal_sd}
\title{Create generic \code{proposal_sd} function for class}
\usage{
proposal_sd(x)
}
\arguments{
\item{x}{An \code{Logistic} object.}
}
\value{
Numeric vector of proposal scales for Metropolis step for regression
coefficients sampling.
}
\description{
Create generic \code{proposal_sd} function for class
}
\examples{
m1 <- Logistic(ndocs = 1)
proposal_sd(m1)
}
