\name{psychlda-package}
\alias{psychlda-package}
\alias{psychlda}
\docType{package}
\title{
\packageTitle{psychlda}
}
\description{
\packageDescription{psychlda}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{psychlda}
\packageIndices{psychlda}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{psychlda}

Maintainer: \packageMaintainer{psychlda}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
