% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychtm.R
\docType{package}
\name{psychtm}
\alias{psychtm}
\title{psychtm: A package for text mining methods for psychological research}
\description{
The \code{psychtm} package provides estimation, summarization, and goodness-of-fit
functions:
}
\section{Model Fitting}{

The workhorse function for Bayesian estimation of topic models is
\code{\link[=gibbs_sldax]{gibbs_sldax()}}.
Similarly, see \code{\link[=gibbs_mlr]{gibbs_mlr()}} and
\code{\link[=gibbs_logistic]{gibbs_logistic()}} to estimate regression models with
continuous and dichotomous outcomes, respectively.
}

\section{Parameter Estimates and Goodness-of-Fit}{

See \code{\link{sldax-summary}} for functions to obtain and summarize parameter
estimates and to compute goodness-of-fit metrics.
}

