% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pwaic_d}
\alias{pwaic_d}
\title{Effective number of parameters for WAIC in sLDA logistic from y_d}
\usage{
pwaic_d(loglike_pred)
}
\arguments{
\item{loglike_pred}{A m x 1 matrix of log-predictive likelihoods.}
}
\description{
Effective number of parameters for WAIC in sLDA logistic from y_d
}
