% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm_cpp}
\alias{rmvnorm_cpp}
\title{Sample from multivariate Gaussian N(\eqn{\mu}, \eqn{\Sigma})}
\usage{
rmvnorm_cpp(n, mu, sigma)
}
\arguments{
\item{n}{The number of samples to draw.}

\item{mu}{The q x 1 mean vector of the distribution (column vector).}

\item{sigma}{The q x q variance-covariance matrix of the distribution.}
}
\value{
A n x q matrix of random draws.
}
\description{
Sample from multivariate Gaussian N(\eqn{\mu}, \eqn{\Sigma})
}
