% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{se_waic<-}
\alias{se_waic<-}
\title{Create generic \verb{se_waic<-} function for class}
\usage{
se_waic(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Numeric standard error of WAIC estimate to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{se_waic<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
se_waic(m1) <- NaN
}
