% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{se_waic}
\alias{se_waic}
\title{Create generic \code{se_waic} function for class}
\usage{
se_waic(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
Numeric standard error for WAIC estimate.
}
\description{
Create generic \code{se_waic} function for class
}
\examples{
m1 <- Model(ndocs = 1)
se_waic(m1)
}
