% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{sigma0<-}
\alias{sigma0<-}
\title{Create generic \verb{sigma0<-} function for class}
\usage{
sigma0(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Numeric covariance matrix of prior for regression coefficients
to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{sigma0<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
sigma0(m1) <- diag(1.0, 2)

}
