% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{sigma0}
\alias{sigma0}
\title{Create generic \code{sigma0} function for class}
\usage{
sigma0(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
Double matrix of prior variances and covariances for regression
coefficients.
}
\description{
Create generic \code{sigma0} function for class
}
\examples{
m1 <- Model(ndocs = 1)
sigma0(m1)
}
