% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{sigma2<-}
\alias{sigma2<-}
\title{Create generic \verb{sigma2<-} function for class}
\usage{
sigma2(x) <- value
}
\arguments{
\item{x}{An \code{Mlr} object.}

\item{value}{Numeric value of residual variance to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{sigma2<-} function for class
}
\examples{
m1 <- Mlr(ndocs = 1)
sigma2(m1) <- 1.0
}
