% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{sigma2}
\alias{sigma2}
\title{Create generic \code{sigma2} function for class}
\usage{
sigma2(x)
}
\arguments{
\item{x}{An \code{Mlr} object.}
}
\value{
Numeric vector of posterior draws of residual variance.
}
\description{
Create generic \code{sigma2} function for class
}
\examples{
m1 <- Mlr(ndocs = 1)
sigma2(m1)
}
