% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_slda}
\alias{sim_slda}
\title{Simulate data from the sLDA model}
\usage{
sim_slda(D, V, N, K, theta, beta, eta, sigma2)
}
\arguments{
\item{D}{The number of documents in the corpus.}

\item{V}{The number of terms in the corpus vocabulary.}

\item{N}{A D x 1 vector of the number of words in each document.}

\item{K}{The number of topics.}

\item{theta}{A D x K matrix of topic proportions (each row must sum to 1).}

\item{beta}{A K x V matrix of word probabilities per topic (each row must
sum to 1).}
}
\description{
Simulate data from the sLDA model
}
