% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-class.R
\docType{class}
\name{Sldalogit-class}
\alias{Sldalogit-class}
\alias{Sldalogit}
\title{An S4 class to represent a sLDA logistic model.}
\description{
An S4 class to represent a sLDA logistic model.
}
\section{Slots}{

\describe{
\item{\code{eta}}{A M x K numeric matrix of draws of topic regression coefficients}

\item{\code{mu0}}{A K x 1 numeric matrix of prior means for eta}

\item{\code{sigma0}}{A K x K numeric prior covariance matrix for eta}

\item{\code{eta_start}}{A K x 1 numeric matrix of starting values for eta}

\item{\code{ntopics}}{The number of topics for the LDA model (default: 2).}

\item{\code{ndocs}}{The number of documents in the corpus.}

\item{\code{nvocab}}{The number of terms in the corpus vocabulary.}

\item{\code{nchain}}{The number of iterations of the Gibbs sampler.}

\item{\code{topics}}{A D x max(N_d) x M numeric array of topic draws. 0 indicates an
unused word index (i.e., the document did not have a word at that index).}

\item{\code{beta}}{A K x V x M numeric array of draws of topic-word probabilities.}

\item{\code{theta}}{A D x K x M numeric array of draws of document-topic
probabilities.}

\item{\code{alpha}}{A numeric prior hyperparameter for theta.}

\item{\code{gamma}}{A numeric prior hyperparameter for beta.}

\item{\code{loglike}}{The log-likelihood (up to an additive constant).}

\item{\code{logpost}}{The log-posterior (up to an additive constant).}

\item{\code{p_eff}}{The effective number of parameters.}

\item{\code{waic}}{WAIC (up to an additive constant) on the deviance scale.}

\item{\code{se_waic}}{Standard error of the WAIC.}

\item{\code{lpd}}{A matrix of iterations x observations of predictive posterior
likelihoods (NOT log-likelihoods).}
}}

