% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{waic<-}
\alias{waic<-}
\title{Create generic \verb{waic<-} function for class}
\usage{
waic(x) <- value
}
\arguments{
\item{x}{An \code{Model} object.}

\item{value}{Numeric value of WAIC to assign to slot.}
}
\value{
None.
}
\description{
Create generic \verb{waic<-} function for class
}
\examples{
m1 <- Model(ndocs = 1)
waic(m1) <- NaN
}
