% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\name{waic}
\alias{waic}
\title{Create generic \code{waic} function for class}
\usage{
waic(x)
}
\arguments{
\item{x}{An \code{Model} object.}
}
\value{
Numeric value of the Watanabe Information Criterion (WAIC).
}
\description{
Create generic \code{waic} function for class
}
\examples{
m1 <- Model(ndocs = 1)
waic(m1)
}
