% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{waic_all}
\alias{waic_all}
\title{Compute WAIC for all outcomes.}
\usage{
waic_all(D, iter, l_pred)
}
\arguments{
\item{D}{The number of documents.}

\item{iter}{The current iteration of the chain.}

\item{l_pred}{A m x D matrix of predictive likelihoods.}
}
\description{
Compute WAIC for all outcomes.
}
