% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{waic_diff}
\alias{waic_diff}
\title{Compute difference (WAIC1 - WAIC2) in WAIC and its SE for two models.}
\usage{
waic_diff(D, m1, m2, l_pred1, l_pred2)
}
\arguments{
\item{D}{The number of documents.}

\item{m1}{The length of the chain for model 1.}

\item{m2}{The length of the chain for model 2.}

\item{l_pred1}{A m x D matrix of predictive likelihoods from model 1.}

\item{l_pred2}{A m x D matrix of predictive likelihoods from model 2.}
}
\description{
Compute difference (WAIC1 - WAIC2) in WAIC and its SE for two models.
}
