% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_topic_word}
\alias{count_topic_word}
\title{Count topic-word co-occurences in corpus (ntopic x nvocab)}
\usage{
count_topic_word(K, V, doc_topic, doc_word)
}
\arguments{
\item{K}{The number of topics.}

\item{V}{The number of terms in the corpus vocabulary.}

\item{doc_topic}{A D x max(\eqn{N_d}) matrix of topic assignments for
the corpus.}

\item{doc_word}{A D x max(\eqn{N_d}) matrix of words for corpus.}
}
\value{
A K x V matrix of topic-word co-occurence counts.
}
\description{
Count topic-word co-occurences in corpus (ntopic x nvocab)
}
