% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eta_logpost_logit}
\alias{eta_logpost_logit}
\title{Compute full conditional log-posterior of eta for logistic sLDA}
\usage{
eta_logpost_logit(zbar, y, eta, mu0, sigma0)
}
\arguments{
\item{zbar}{A D x K matrix with row d containing the mean number of draws of
topics \eqn{z_1, \ldots, z_K} in document \eqn{d} where each row sums to
1.}

\item{y}{A D x 1 vector of the outcome variable for each document.}

\item{mu0}{A K x 1 vector of prior means for the regression coefficients.}

\item{sigma0}{A K x K prior variance-covariance matrix for the regression
coefficients.}
}
\description{
Compute full conditional log-posterior of eta for logistic sLDA
}
