% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{get_topwords,Lda,numeric,character-method}
\alias{get_topwords,Lda,numeric,character-method}
\title{Return most probable words for document}
\usage{
\S4method{get_topwords}{Lda,numeric,character}(mcmc_fit, n_words, vocab,
  burn = 0, thin = 1, stat = "mean")
}
\description{
Return most probable words for document
}
