% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{get_topwords,matrix,numeric,character-method}
\alias{get_topwords,matrix,numeric,character-method}
\title{Compute the most probable words in a topic}
\usage{
\S4method{get_topwords}{matrix,numeric,character}(beta_, nwords, vocab,
  method = "termscore")
}
\arguments{
\item{beta_}{A \eqn{K} x \eqn{V} matrix of \eqn{V} vocabulary probabilities
for each of \eqn{K} topics.}

\item{nwords}{The number of words to retrieve.}

\item{vocab}{A character vector containing the vocabulary.}

\item{method}{If "termscore", use term scores (similar to tf-idf). If "prob",
use probabilities. Default: "termscore".}
}
\description{
Compute the most probable words in a topic
}
