% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{get_zbar,Lda-method}
\alias{get_zbar,Lda-method}
\title{Generic function to retrieve the empirical topic proportions}
\usage{
\S4method{get_zbar}{Lda}(mcmc_fit, burn = 0, thin = 1)
}
\arguments{
\item{mcmc_fit}{An Lda object.}

\item{burn}{The number of draws to discard as a burn-in period. Default: 0.}

\item{thin}{The number of draws to skip as a thinning period. Default: 1 (no thinning).}
}
\description{
Generic function to retrieve the empirical topic proportions
}
