% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{get_zbar,Lda-method}
\alias{get_zbar,Lda-method}
\title{Return empirical topic proportions}
\usage{
\S4method{get_zbar}{Lda}(mcmc_fit, burn = 0, thin = 1)
}
\description{
Compute empirical topic proportions (zbar) from \code{@topics} in an Lda
object
}
