% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\docType{methods}
\name{gg_coef,Sldalogit-method}
\alias{gg_coef,Sldalogit-method}
\title{Generic function to plot the regression coefficients for sLDAX models}
\usage{
\S4method{gg_coef}{Sldalogit}(mcmc_fit, burn = 0, thin = 1,
  stat = "mean", errorbw = 0.5)
}
\arguments{
\item{mcmc_fit}{An Sldalogit object.}

\item{burn}{The number of draws to discard as a burn-in period. Default: 0.}

\item{thin}{The number of draws to skip as a thinning period. Default: 1 (no thinning).}

\item{stat}{The summary statistic to use on the posterior draws. Default: \code{"mean"}.}

\item{errorbw}{Controls the width of the +/- 2 posterior standard error bars.}
}
\description{
Generic function to plot the regression coefficients for sLDAX models
}
