% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_logistic_cpp}
\alias{gibbs_logistic_cpp}
\title{Collapsed Gibbs sampler for logistic regression}
\usage{
gibbs_logistic_cpp(m, burn, thin, y, x, mu0, sigma0, eta_start,
  proposal_sd, verbose = FALSE, display_progress = FALSE)
}
\arguments{
\item{m}{The number of iterations to run the Gibbs sampler.}

\item{burn}{The number of iterations to discard as the burn-in period.}

\item{y}{A D x 1 vector of binary outcomes (0/1) to be predicted.}

\item{x}{A D x p matrix of additional predictors (no column of 1s for
intercept).}

\item{mu0}{A (p + 1) x 1 mean vector for the prior on the regression coefficients.}

\item{sigma0}{A (p + 1) x (p + 1) variance-covariance matrix for the prior
on the regression coefficients.}

\item{eta_start}{A (p + 1) x 1 vector of starting values for the
regression coefficients.}

\item{proposal_sd}{The proposal standard deviation for drawing the
regression coefficients, N(0, proposal_sd) (default: 2.38, ..., 2.38).}

\item{verbose}{Should parameter draws be output during sampling? (default:
\code{FALSE}).}

\item{display_progress}{Should percent progress of sampler be displayed
(default: \code{FALSE}). Recommended that only one of \code{verbose} and
\code{display_progress} be set to \code{TRUE} at any given time.}
}
\value{
An object of class \code{Logistic}.
}
\description{
Collapsed Gibbs sampler for logistic regression
}
\seealso{
Other Gibbs sampler: \code{\link{gibbs_logistic}},
  \code{\link{gibbs_mlr_cpp}}, \code{\link{gibbs_mlr}},
  \code{\link{gibbs_sldax}}
}
\concept{Gibbs sampler}
