% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sldax_cpp}
\alias{gibbs_sldax_cpp}
\alias{.gibbs_sldax_cpp}
\title{Collapsed Gibbs sampler for the sLDA-X model}
\usage{
.gibbs_sldax_cpp(
  docs,
  V,
  m,
  burn,
  thin,
  K,
  model,
  y,
  x,
  mu0,
  sigma0,
  a0,
  b0,
  eta_start,
  proposal_sd,
  interaction_xcol = -1L,
  alpha_ = 1,
  gamma_ = 1,
  constrain_eta = FALSE,
  sample_beta = TRUE,
  sample_theta = TRUE,
  return_assignments = FALSE,
  verbose = FALSE,
  display_progress = FALSE
)
}
\arguments{
\item{docs}{A D x max(\eqn{N_d}) matrix of word indices for all documents.}

\item{V}{The number of unique terms in the vocabulary.}

\item{m}{The number of iterations to run the Gibbs sampler.}

\item{burn}{The number of iterations to discard as the burn-in period.}

\item{thin}{The period of iterations to keep after the burn-in period
(default: 1).}

\item{K}{The number of topics.}

\item{model}{An integer denoting the type of model to fit.}

\item{y}{A D x 1 vector of binary outcomes (0/1) to be predicted.}

\item{x}{A D x p matrix of additional predictors (no column of 1s for
intercept).}

\item{mu0}{A K x 1 mean vector for the prior on the regression coefficients.}

\item{sigma0}{A (K + p + 1) x (K + p + 1) variance-covariance matrix for the
prior on the regression coefficients. The first p + 1 columns/rows
correspond to predictors in X, while the last K columns/rows correspond to
the K topic means.}

\item{a0}{The shape parameter for the prior on sigma2.}

\item{b0}{The scale parameter for the prior on sigma2.}

\item{eta_start}{A (K + p) x 1 vector of starting values for the
regression coefficients. The first p elements correspond to predictors
in X, while the last K elements correspond to the K topic means.}

\item{proposal_sd}{The proposal standard deviation for drawing the
regression coefficients, N(0, proposal_sd) (default: 0.2).}

\item{interaction_xcol}{The column number of the design matrix for the
additional predictors for which an interaction with the \eqn{K} topics is
desired (default: \eqn{-1L}, no interaction). Currently only supports a
single continuous predictor or a two-category categorical predictor
represented as a single dummy-coded column.}

\item{alpha_}{The hyper-parameter for the prior on the topic proportions
(default: 1.0).}

\item{gamma_}{The hyper-parameter for the prior on the topic-specific
vocabulary probabilities (default: 1.0).}

\item{constrain_eta}{A logical (default = \code{false}): If \code{true}, the
regression coefficients will be constrained so that they are in descending
order; if \code{FALSE}, no constraints will be applied.}

\item{sample_beta}{A logical (default = \code{true}): If \code{true}, the
topic-vocabulary distributions are sampled from their full conditional
distribution.}

\item{sample_theta}{A logical (default = \code{true}): If \code{true}, the
topic proportions are sampled from their full conditional distribution.}

\item{return_assignments}{A logical (default = \code{false}): If
\code{true}, returns an N x \eqn{max N_d} x M array of topic assignments
in slot @topics. CAUTION: this can be memory-intensive.}

\item{verbose}{Should parameter draws be output during sampling? (default:
\code{False}).}

\item{display_progress}{Should percent progress of sampler be displayed
(default: \code{false}). Recommended that only one of \code{verbose} and
\code{display_progress} be set to \code{true} at any given time.}
}
\description{
Collapsed Gibbs sampler for the sLDA-X model
}
