% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-class.R
\docType{class}
\name{Lda-class}
\alias{Lda-class}
\alias{Lda}
\title{An S4 class to represent a LDA model.}
\description{
An S4 class to represent a LDA model.
}
\section{Slots}{

\describe{
\item{\code{ntopics}}{The number of topics for the LDA model (default: 2).}

\item{\code{ndocs}}{The number of documents in the corpus.}

\item{\code{nvocab}}{The number of terms in the corpus vocabulary.}

\item{\code{nchain}}{The number of iterations of the Gibbs sampler.}

\item{\code{topics}}{A D x max(N_d) x M numeric array of topic draws. 0 indicates an
unused word index (i.e., the document did not have a word at that index).}

\item{\code{alpha}}{A numeric prior hyperparameter for theta.}

\item{\code{gamma}}{A numeric prior hyperparameter for beta.}

\item{\code{loglike}}{The log-likelihood (up to an additive constant).}

\item{\code{logpost}}{The log-posterior (up to an additive constant).}
}}

