% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R, R/aaa-generics.R
\docType{class}
\name{Logistic-class}
\alias{Logistic-class}
\alias{proposal_sd}
\alias{proposal_sd,Logistic-method}
\alias{proposal_sd<-}
\alias{proposal_sd<-,Logistic-method}
\alias{Logistic}
\title{S4 class for a logistic regression model that inherits from \linkS4class{Model}}
\usage{
proposal_sd(x)

\S4method{proposal_sd}{Logistic}(x)

proposal_sd(x) <- value

\S4method{proposal_sd}{Logistic}(x) <- value

Logistic(proposal_sd = NaN, ...)
}
\arguments{
\item{x}{An \code{Logistic} object.}

\item{value}{A value to assign to a slot for \code{x}}

\item{proposal_sd}{A vector of p + 1 proposal scales/standard deviations for
sampling of p + 1 regression coefficients by Metropolis-Hastings.}

\item{...}{additional arguments to be passed to the low level regression
fitting functions (see below).}
}
\description{
S4 class for a logistic regression model that inherits from \linkS4class{Model}

Create generic \code{proposal_sd} function for class

Slot \verb{@proposal_sd} generic accessor

Create generic \verb{proposal_sd<-} function for class

Logistic helper function (constructor) for \code{Logistic} class
}
\section{Slots}{

\describe{
\item{\code{proposal_sd}}{A vector of p + 1 proposal scales/standard deviations for
sampling of p + 1 regression coefficients by Metropolis-Hastings.}
}}

\keyword{classes}
