% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-class.R
\docType{class}
\name{Logistic-class}
\alias{Logistic-class}
\alias{Logistic}
\title{An S4 class to represent a logistic regression model.}
\description{
An S4 class to represent a logistic regression model.
}
\section{Slots}{

\describe{
\item{\code{eta}}{A M x (p + 1) numeric matrix of draws of topic regression coefficients}

\item{\code{mu0}}{A (p + 1) x 1 numeric matrix of prior means for eta}

\item{\code{sigma0}}{A (p + 1) x (p + 1) numeric prior covariance matrix for eta}

\item{\code{eta_start}}{A (p + 1) x 1 numeric matrix of starting values for eta}

\item{\code{proposal_sd}}{A K x 1 vector of proposal scales for Metropolis-Hastings
sampling of eta.}

\item{\code{ndocs}}{The number of documents in the corpus.}

\item{\code{nchain}}{The number of iterations of the Gibbs sampler.}

\item{\code{loglike}}{The log-likelihood (up to an additive constant).}

\item{\code{logpost}}{The log-posterior (up to an additive constant).}

\item{\code{p_eff}}{The effective number of parameters.}

\item{\code{waic}}{WAIC (up to an additive constant) on the deviance scale.}

\item{\code{se_waic}}{Standard error of the WAIC.}

\item{\code{lpd}}{A matrix of iterations x observations of predictive posterior
likelihoods (NOT log-likelihoods).}
}}

