% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R, R/aaa-generics.R
\docType{class}
\name{Logistic-class}
\alias{Logistic-class}
\alias{proposal_sd,Logistic-method}
\alias{proposal_sd<-,Logistic-method}
\alias{Logistic}
\title{S4 class for a logistic regression model that inherits from \linkS4class{Model}}
\usage{
\S4method{proposal_sd}{Logistic}(x)

\S4method{proposal_sd}{Logistic}(x) <- value

Logistic(proposal_sd = NaN, ...)
}
\arguments{
\item{x}{An \code{Logistic} object.}

\item{value}{A value to assign to a slot for \code{x}}

\item{proposal_sd}{A vector of p + 1 proposal scales/standard deviations for
sampling of p + 1 regression coefficients by Metropolis-Hastings.}

\item{...}{additional arguments to be passed to the low level regression
fitting functions (see below).}
}
\value{
A \linkS4class{Logistic} object.
}
\description{
S4 class for a logistic regression model that inherits from \linkS4class{Model}

Helper function (constructor) for \code{Logistic} class
}
\section{Slots}{

\describe{
\item{\code{proposal_sd}}{A vector of p + 1 proposal scales/standard deviations for
sampling of p + 1 regression coefficients by Metropolis-Hastings.}
}}

\examples{
m1 <- Logistic(ndocs = 1)
print(m1)

}
\keyword{classes}
