% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-class.R
\name{Logistic}
\alias{Logistic}
\title{An S4 class to represent a logistic regression model.}
\usage{
Logistic(proposal_sd = NaN, ...)

Logistic(proposal_sd = NaN, ...)
}
\arguments{
\item{proposal_sd}{A (p + 1) x 1 vector of proposal scales for
Metropolis-Hastings sampling of eta.}
}
\description{
An S4 class to represent a logistic regression model.

Helper function (constructor) for Logistic class.
}
\section{Slots}{

\describe{
\item{\code{proposal_sd}}{A (p + 1) x 1 vector of proposal scales for
Metropolis-Hastings sampling of eta.}
}}

