% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R, R/aaa-generics.R
\name{Logistic}
\alias{Logistic}
\alias{proposal_sd}
\alias{proposal_sd,Logistic-method}
\alias{proposal_sd<-}
\alias{proposal_sd<-,Logistic-method}
\title{Logistic class}
\usage{
Logistic(proposal_sd = NaN, ...)

proposal_sd(x)

\S4method{proposal_sd}{Logistic}(x)

proposal_sd(x) <- value

\S4method{proposal_sd}{Logistic}(x) <- value

Logistic(proposal_sd = NaN, ...)
}
\arguments{
\item{proposal_sd}{A vector of p + 1 proposal scales/standard deviations for
sampling of p + 1 regression coefficients by Metropolis-Hastings.}

\item{...}{additional arguments to be passed to the low level regression
fitting functions (see below).}

\item{x}{An \code{Logistic} object.}

\item{value}{A value to assign to a slot for \code{x}}
}
\description{
S4 class for a logistic regression model. Inherits from \linkS4class{Model}.
}
\section{Slots}{

\describe{
\item{\code{proposal_sd}}{A vector of p + 1 proposal scales/standard deviations for
sampling of p + 1 regression coefficients by Metropolis-Hastings.}
}}

