% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr-class.R
\name{Mlr}
\alias{Mlr}
\title{An S4 class to represent a regression model.}
\usage{
Mlr(a0 = 0.001, b0 = 0.001, sigma2 = NaN, ...)

Mlr(a0 = 0.001, b0 = 0.001, sigma2 = NaN, ...)
}
\arguments{
\item{a0}{A prior shape hyperparameter for sigma2.}

\item{b0}{A prior rate hyperparameter for sigma2.}

\item{sigma2}{A nchain x 1 numeric vector of draws of the residual variance.}
}
\description{
An S4 class to represent a regression model.

Helper function (constructor) for Mlr class.
}
\section{Slots}{

\describe{
\item{\code{a0}}{A prior shape hyperparameter for sigma2.}

\item{\code{b0}}{A prior rate hyperparameter for sigma2.}

\item{\code{sigma2}}{A nchain x 1 numeric vector of draws of the residual variance.}
}}

