% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{post_pred_sldax_logit}
\alias{post_pred_sldax_logit}
\title{Posterior predictive likelihood for sLDA-X logistic}
\usage{
post_pred_sldax_logit(x, zbar, eta)
}
\arguments{
\item{x}{A D x p matrix of additional predictors.}

\item{zbar}{A D x K matrix with row \eqn{d} containing the mean number of
draws of topics \eqn{z_1, \ldots, z_K} in document \eqn{d} where each row
sums to 1.}

\item{eta}{A (p + K) x 1 vector of regression coefficients.}
}
\description{
Posterior predictive likelihood for sLDA-X logistic
}
