% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pwaic_d}
\alias{pwaic_d}
\title{Contribution to effective number of parameters for WAIC from observation y_d}
\usage{
pwaic_d(like_pred)
}
\arguments{
\item{like_pred}{A m x 1 vector of predictive likelihoods.}
}
\description{
Contribution to effective number of parameters for WAIC from observation y_d
}
