% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-class.R
\docType{class}
\name{Slda-class}
\alias{Slda-class}
\alias{Slda}
\title{An S4 class to represent a sLDA model.}
\description{
An S4 class to represent a sLDA model.
}
\section{Slots}{

\describe{
\item{\code{eta}}{A M x K numeric matrix of draws of topic regression coefficients}

\item{\code{sigma2}}{A M x 1 numeric vector of draws of residual variance}

\item{\code{beta}}{A K x V x M numeric array of draws of topic-word probabilities}

\item{\code{theta}}{A D x K x M numeric array of draws of document-topic
probabilities}

\item{\code{mu0}}{A K x 1 numeric matrix of prior means for eta}

\item{\code{sigma0}}{A K x K numeric prior covariance matrix for eta}

\item{\code{alpha}}{A numeric prior hyperparameter for theta}

\item{\code{gamma}}{A numeric prior hyperparameter for beta}

\item{\code{a0}}{A numeric prior shape hyperparameter for sigma2}

\item{\code{b0}}{A numeric prior rate hyperparameter for sigma2}

\item{\code{eta_start}}{A K x 1 numeric matrix of starting values for eta}
}}

