% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-generic-functions.R
\name{get_toptopics}
\alias{get_toptopics}
\alias{get_topwords}
\title{Generic function to retrieve the top topics for each document}
\usage{
get_toptopics(mcmc_fit, n_topics, burn = 0, thin = 1, stat = "mean")

get_topwords(mcmc_fit, n_words, vocab, burn = 0, thin = 1,
  stat = "mean")
}
\arguments{
\item{mcmc_fit}{An SLda object.}

\item{n_topics}{The number of topics to retrieve.}

\item{burn}{The number of draws to discard as a burn-in period. Default: 0.}

\item{thin}{The number of draws to skip as a thinning period. Default: 1 (no thinning).}

\item{stat}{The summary statistic to use on the posterior draws. Default: mean.}

\item{n_words}{The number of words to retrieve.}

\item{vocab}{A character vector containing the vocabulary.}

\item{mcmc_fit}{An SLda object.}

\item{burn}{The number of draws to discard as a burn-in period. Default: 0.}

\item{thin}{The number of draws to skip as a thinning period. Default: 1 (no thinning).}

\item{stat}{The summary statistic to use on the posterior draws. Default: mean.}
}
\value{
A tibble containing \code{doc}, \code{topic}, and \code{prob}.

A tibble containing \code{topic}, \code{word}, and \code{prob}
}
\description{
Description of get_toptopics here.

Description of get_topwords here.
}
