% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sldax-class.R
\name{Sldax}
\alias{Sldax}
\title{An S4 class to represent a sLDAX general model.}
\usage{
Sldax(nvocab, topics, ntopics = 2, alpha = 0.1, gamma = 1.01, ...)

Sldax(nvocab, topics, ntopics = 2, alpha = 0.1, gamma = 1.01, ...)
}
\arguments{
\item{nvocab}{The number of terms in the corpus vocabulary.}

\item{topics}{A D x max(N_d) x M numeric array of topic draws. 0 indicates an
unused word index (i.e., the document did not have a word at that index).}

\item{ntopics}{The number of topics for the LDA model (default: 2).}

\item{alpha}{A numeric prior hyperparameter for theta.}

\item{gamma}{A numeric prior hyperparameter for beta.}
}
\description{
An S4 class to represent a sLDAX general model.

Helper function (constructor) for Sldax class.
}
\section{Slots}{

\describe{
\item{\code{nvocab}}{The number of terms in the corpus vocabulary.}

\item{\code{ntopics}}{The number of topics for the LDA model (default: 2).}

\item{\code{alpha}}{A numeric prior hyperparameter for theta.}

\item{\code{gamma}}{A numeric prior hyperparameter for beta.}

\item{\code{topics}}{A D x max(N_d) x M numeric array of topic draws. 0 indicates an
unused word index (i.e., the document did not have a word at that index).}
}}

