% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slda-results.R
\name{term_score}
\alias{term_score}
\title{Compute term-scores for each word-topic pair (p. 75, Srivasta & Sahami, 2009)}
\usage{
term_score(beta_)
}
\arguments{
\item{beta_}{A K x V matrix of V vocabulary probabilities for each of K topics.}
}
\value{
A K x V matrix of term-scores (comparable to tf-idf).
}
\description{
Compute term-scores for each word-topic pair (p. 75, Srivasta & Sahami, 2009)
}
