% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{waic_all}
\alias{waic_all}
\title{Compute WAIC for all outcomes.}
\usage{
waic_all(iter, l_pred)
}
\arguments{
\item{iter}{The current iteration of the chain.}

\item{l_pred}{A m x D matrix of predictive likelihoods (NOT log-likelihoods).}
}
\value{
Vector of (1) WAIC for model, (2) standard error for WAIC, and (3)
  the effective number of parameters.
}
\description{
Compute WAIC for all outcomes.
}
