% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{waic_d}
\alias{waic_d}
\title{WAIC for observation y_d}
\arguments{
\item{like_pred}{A m x 1 vector of predictive likelihoods (NOT log-likelihoods) for y_d.}

\item{p_effd}{The contribution to the effective number of parameters from
obs y_d.}
}
\value{
WAIC contribution for observation d (on deviance scale).
}
\description{
WAIC for observation y_d
}
